+++ src/main/kotlin/dev/aaronhowser/mods/ariadnesthread/datagen/ModDataGen.kt
@@ -24,6 +24,8 @@ object ModDataGen {
         val lookupProvider: CompletableFuture<HolderLookup.Provider> = event.lookupProvider
 
         val languageProvider = generator.addProvider(event.includeClient(), ModLanguageProvider(output))
+
+        val recipeProvider = generator.addProvider(event.includeServer(), ModRecipeProvider(output, lookupProvider))
     }
 
 }