+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 
 - Update NeoForge to 21.0.52-beta from 21.0.14-beta
 - Removed LocationItemComponent, HistoryItemComponent now has a List<BlockPos> instead. As a consequence, it no longer stores the exact float location. This is better for performance.
+- Changed the text in the tooltip